requirements.txt
streamlit
openai-whisper
torch
deep-translator
gTTS
easyocr
opencv-python-headless
Pillow
numpy
pydub
streamlit-audiorec
SpeechRecognition
audio-recorder-streamlit
